Added NEL to John Brelsford places
I encountered one error: dvp_078_place_74 should be tagged as an organization, not a place, and should have the QID of Q592576. I am not sure how to change the tagging of this entity because that would change all of the IDs of each place tag and each organization tag.